Install only r-base-core
r-recommended was bringing in a lot of apt packages
which eventually cause problems as they want a new version
of R, and we want to try pin to 4.0. This is described in
https://github.com/jupyterhub/repo2docker/issues/1116.

I verified that the packages no longer being installed by
apt are still installed via install.packages, primarily as dependencies
of packages we are installing.